Create Week2_Proj1
The Pima Indians diabetes data set (pima-indians-diabetes.xlsx) is a data set used to diagnostically
predict whether or not a patient has diabetes, based on certain diagnostic measurements included in the dataset.
All patients here are females at least 21 years old of Pima Indian heritage. The dataset consists of M = 8 attributes
and one target variable, Outcome (1 represents diabetes, 0 represents no diabetes). The 8 attributes include
Pregnancies, Glucose, BloodPressure, BMI, insulin level, age, and so on. There are N=768 data samples.
Randomly select n samples from the “diabetes” class and n samples from the “no diabetes” class, and use them
as the training samples. The remaining data samples are the test samples. Build a linear regression model with the
training set, and test your model on the test samples to predict whether or not a test patient has diabetes or not.
Assume the predicted outcome of a test sample is 𝑡̂, if 𝑡̂ ≥ 0.5 (closer to 1), classify it as “diabetes”; if 𝑡̂ < 0.5
(closer to 0), classify it as “no diabetes”. Run 1000 independent experiments, and calculate the prediction accuracy
rate as 𝑡ℎ𝑒 𝑛𝑢𝑚𝑏𝑒𝑟 𝑜𝑓 𝑐𝑜𝑟𝑟𝑒𝑐𝑡 𝑝𝑟𝑒𝑑𝑖𝑐𝑡𝑖𝑜𝑛𝑠 / 𝑡ℎ𝑒 𝑡𝑜𝑡𝑎𝑙 𝑛𝑢𝑚𝑏𝑒𝑟 𝑜𝑓 𝑡𝑒𝑠𝑡 𝑐𝑎𝑠𝑒𝑠 %. Let n=20, 40, 60, 80, 100, plot the accuracy rate versus n.